nlp_label_quality/requirements.txt
setuptools~=49.2.1
future~=0.18.2
Pillow~=8.0.1
numpy~=1.20.1
gensim~=3.8.3
spacy~=2.3.2
nltk~=3.5
pm4py~=2.2.1
pandas~=1.2.3
python-dateutil~=2.8.1